Sneat/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1
